Update example 1 so it can be compiled for desktop
I updated the code so that it can be compiled with the new
desktop_build.bat script in order to create a version which can be
played offline (and debugged easier).  I also added some quailty of life
scripts to help generate local hosts and open the webpage that has the
host.